+++ Cache simulator/RAM.h
@@ -19,38 +19,38 @@ private:
 public:	
 
 	// for statistics
-	long int readReq;
-	long int writeReq;
-	long int missCounter;
-	long int wrongAdd;
+	long long readReq;
+	long long writeReq;
+	long long missCounter;
+	long long wrongAdd;
 
 	int size;
-	long int filled;												// It is number of enmtry. It mustn't be exceed RAM_size from param.h	
+	long long filled;												// It is number of enmtry. It mustn't be exceed RAM_size from param.h	
 			
 	list <Chunk*> LRU;												// LRU structure for eviction
 	
-																	/*Param long int: address of chunk.	
+																	/*Param long long: address of chunk.	
 	Param bool: 
 	false = this lba haven't been read after adding
 	true = this lba have been read after adding*/
-	unordered_map <long int, pair< list<Chunk*>::iterator, bool>> HashTable;		// Structure for searches
+	unordered_map <long long, pair< list<Chunk*>::iterator, bool>> HashTable;		// Structure for searches
 	
 	// 100% size
 	RAM();
 	
 	void evict();
-	void add(long int num);
-	void update(long int num);
+	void add(long long num);
+	void update(long long num);
 
-	bool read(long int addr);
+	bool read(long long addr);
 
-	void write(long int addr);
+	void write(long long addr);
 
-	bool exist(long int addr);
+	bool exist(long long addr);
 	
-	void remove(long int addr);
+	void remove(long long addr);
 
-	void saveStatistics(long int lba_counter, string traceName);
+	void saveStatistics(long long lba_counter, string traceName);
 	
 	void cleanAndResize(int new_size);
 };
